Folded MapServiceContextSupport into MapServiceContext Folded theAbstractMapServiceContextSupport into MapServiceContextImpl
