+++ hazelcast/src/main/java/com/hazelcast/map/impl/MapServiceContextImpl.java
@@ -19,12 +19,17 @@ package com.hazelcast.map.impl;
 import com.hazelcast.config.Config;
 import com.hazelcast.config.MapConfig;
 import com.hazelcast.core.HazelcastException;
+import com.hazelcast.core.PartitioningStrategy;
+import com.hazelcast.map.MapInterceptor;
 import com.hazelcast.map.impl.eviction.EvictionOperator;
 import com.hazelcast.map.impl.eviction.ExpirationManager;
+import com.hazelcast.map.listener.MapPartitionLostListener;
 import com.hazelcast.map.merge.MergePolicyProvider;
 import com.hazelcast.nio.Address;
 import com.hazelcast.nio.serialization.Data;
 import com.hazelcast.partition.InternalPartitionService;
+import com.hazelcast.spi.EventFilter;
+import com.hazelcast.spi.EventRegistration;
 import com.hazelcast.spi.NodeEngine;
 import com.hazelcast.util.ConcurrencyUtil;
 import com.hazelcast.util.ConstructorFunction;
@@ -32,17 +37,22 @@ import com.hazelcast.util.ConstructorFunction;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedHashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 
+import static com.hazelcast.map.impl.ListenerAdapters.createListenerAdapter;
+import static com.hazelcast.map.impl.MapService.SERVICE_NAME;
+
 /**
  * Default implementation of map service context.
  */
-class DefaultMapServiceContext extends AbstractMapServiceContextSupport {
+class MapServiceContextImpl implements MapServiceContext {
 
+    private final NodeEngine nodeEngine;
     private final PartitionContainer[] partitionContainers;
     private final ConcurrentMap<String, MapContainer> mapContainers;
     private final AtomicReference<Collection<Integer>> ownedPartitions;
@@ -71,8 +81,8 @@ class DefaultMapServiceContext extends AbstractMapServiceContextSupport {
     private EvictionOperator evictionOperator;
     private MapService mapService;
 
-    public DefaultMapServiceContext(NodeEngine nodeEngine) {
-        super(nodeEngine);
+    public MapServiceContextImpl(NodeEngine nodeEngine) {
+        this.nodeEngine = nodeEngine;
         int partitionCount = nodeEngine.getPartitionService().getPartitionCount();
         this.partitionContainers = new PartitionContainer[partitionCount];
         this.mapContainers = new ConcurrentHashMap<String, MapContainer>();
@@ -274,5 +284,199 @@ class DefaultMapServiceContext extends AbstractMapServiceContextSupport {
     @Override
     public void setEvictionOperator(EvictionOperator evictionOperator) {
         this.evictionOperator = evictionOperator;
+    }
+
+    @Override
+    public Object toObject(Object data) {
+        if (data == null) {
+            return null;
+        }
+        if (data instanceof Data) {
+            return nodeEngine.toObject(data);
+        } else {
+            return data;
+        }
+    }
+
+    @Override
+    public Data toData(Object object, PartitioningStrategy partitionStrategy) {
+        if (object == null) {
+            return null;
+        }
+        if (object instanceof Data) {
+            return (Data) object;
+        } else {
+            return nodeEngine.getSerializationService().toData(object, partitionStrategy);
+        }
+    }
+
+    @Override
+    public Data toData(Object object) {
+        if (object == null) {
+            return null;
+        }
+        if (object instanceof Data) {
+            return (Data) object;
+        } else {
+            return nodeEngine.getSerializationService().toData(object);
+        }
+    }
+
+    @Override
+    public boolean compare(String mapName, Object value1, Object value2) {
+        if (value1 == null && value2 == null) {
+            return true;
+        }
+        if (value1 == null) {
+            return false;
+        }
+        if (value2 == null) {
+            return false;
+        }
+        final MapContainer mapContainer = getMapContainer(mapName);
+        return mapContainer.getRecordFactory().isEquals(value1, value2);
+    }
+
+    @Override
+    public void interceptAfterGet(String mapName, Object value) {
+        List<MapInterceptor> interceptors = getMapContainer(mapName).getInterceptors();
+        if (!interceptors.isEmpty()) {
+            value = toObject(value);
+            for (MapInterceptor interceptor : interceptors) {
+                interceptor.afterGet(value);
+            }
+        }
+    }
+
+    @Override
+    public Object interceptPut(String mapName, Object oldValue, Object newValue) {
+        List<MapInterceptor> interceptors = getMapContainer(mapName).getInterceptors();
+        Object result = null;
+        if (!interceptors.isEmpty()) {
+            result = toObject(newValue);
+            oldValue = toObject(oldValue);
+            for (MapInterceptor interceptor : interceptors) {
+                Object temp = interceptor.interceptPut(oldValue, result);
+                if (temp != null) {
+                    result = temp;
+                }
+            }
+        }
+        return result == null ? newValue : result;
+    }
+
+    @Override
+    public void interceptAfterPut(String mapName, Object newValue) {
+        List<MapInterceptor> interceptors = getMapContainer(mapName).getInterceptors();
+        if (!interceptors.isEmpty()) {
+            newValue = toObject(newValue);
+            for (MapInterceptor interceptor : interceptors) {
+                interceptor.afterPut(newValue);
+            }
+        }
+    }
+
+    @Override
+    public Object interceptRemove(String mapName, Object value) {
+        List<MapInterceptor> interceptors = getMapContainer(mapName).getInterceptors();
+        Object result = null;
+        if (!interceptors.isEmpty()) {
+            result = toObject(value);
+            for (MapInterceptor interceptor : interceptors) {
+                Object temp = interceptor.interceptRemove(result);
+                if (temp != null) {
+                    result = temp;
+                }
+            }
+        }
+        return result == null ? value : result;
+    }
+
+    @Override
+    public void interceptAfterRemove(String mapName, Object value) {
+        List<MapInterceptor> interceptors = getMapContainer(mapName).getInterceptors();
+        if (!interceptors.isEmpty()) {
+            for (MapInterceptor interceptor : interceptors) {
+                value = toObject(value);
+                interceptor.afterRemove(value);
+            }
+        }
+    }
+
+    @Override
+    public void addInterceptor(String id, String mapName, MapInterceptor interceptor) {
+        MapContainer mapContainer = getMapContainer(mapName);
+        mapContainer.addInterceptor(id, interceptor);
+    }
+
+
+    @Override
+    public String generateInterceptorId(String mapName, MapInterceptor interceptor) {
+        return interceptor.getClass().getName() + interceptor.hashCode();
+    }
+
+    @Override
+    public void removeInterceptor(String mapName, String id) {
+        getMapContainer(mapName).removeInterceptor(id);
+    }
+
+    // todo interceptors should get a wrapped object which includes the serialized version
+    @Override
+    public Object interceptGet(String mapName, Object value) {
+        List<MapInterceptor> interceptors = getMapContainer(mapName).getInterceptors();
+        Object result = null;
+        if (!interceptors.isEmpty()) {
+            result = toObject(value);
+            for (MapInterceptor interceptor : interceptors) {
+                Object temp = interceptor.interceptGet(result);
+                if (temp != null) {
+                    result = temp;
+                }
+            }
+        }
+        return result == null ? value : result;
+    }
+
+    @Override
+    public String addLocalEventListener(Object listener, String mapName) {
+        ListenerAdapter listenerAdaptor = createListenerAdapter(listener);
+        EventRegistration registration = nodeEngine.getEventService().
+                registerLocalListener(SERVICE_NAME, mapName, listenerAdaptor);
+        return registration.getId();
+    }
+
+    @Override
+    public String addLocalEventListener(Object listener, EventFilter eventFilter, String mapName) {
+        ListenerAdapter listenerAdaptor = createListenerAdapter(listener);
+        EventRegistration registration = nodeEngine.getEventService().
+                registerLocalListener(SERVICE_NAME, mapName, eventFilter, listenerAdaptor);
+        return registration.getId();
+    }
+
+    @Override
+    public String addEventListener(Object listener, EventFilter eventFilter, String mapName) {
+        ListenerAdapter listenerAdaptor = createListenerAdapter(listener);
+        EventRegistration registration = nodeEngine.getEventService().
+                registerListener(SERVICE_NAME, mapName, eventFilter, listenerAdaptor);
+        return registration.getId();
+    }
+
+    @Override
+    public String addPartitionLostListener(MapPartitionLostListener listener, String mapName) {
+        final ListenerAdapter listenerAdapter = new InternalMapPartitionLostListenerAdapter(listener);
+        final EventFilter filter = new MapPartitionLostEventFilter();
+        final EventRegistration registration = nodeEngine.getEventService().registerListener(SERVICE_NAME, mapName, filter,
+                listenerAdapter);
+        return registration.getId();
+    }
+
+    @Override
+    public boolean removeEventListener(String mapName, String registrationId) {
+        return nodeEngine.getEventService().deregisterListener(SERVICE_NAME, mapName, registrationId);
+    }
+
+    @Override
+    public boolean removePartitionLostListener(String mapName, String registrationId) {
+        return nodeEngine.getEventService().deregisterListener(SERVICE_NAME, mapName, registrationId);
     }
 }
